% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruteforce.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force method for knapsack problem.}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{is a data.frame containing two columns, w and v.}

\item{W}{is the capacity of the knapsack and is an integer greater than 0.}
}
\value{
Returns a list containing the maximum value of combination of the objects and the objects in the combination.
}
\description{
Brute force method for knapsack problem.
}
\examples{
#'suppressWarnings(RNGversion(min(as.character(getRversion()),"3.5.3")))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <-
 data.frame(
   w=sample(1:4000, size = n, replace = TRUE),
   v=runif(n = n, 0, 10000)
 )

brute_force_knapsack(knapsack_objects[1:8,], W = 3500)
}
