% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic programming solution of knapsack problem}
\usage{
knapsack_dynamic(x, W, fast = FALSE)
}
\arguments{
\item{x}{x is a data.frame containing two columns, w and v.}

\item{W}{W is the capacity of the knapsack and is an integer greater than 0.}

\item{fast}{is a logical argument. If TRUE, Rcpp code is run speeding up the calculations.}
}
\value{
Returns a list containing the maximum value of combination of the objects and the objects in the combination.
}
\description{
Dynamic programming solution of knapsack problem
}
\examples{


suppressWarnings(RNGversion(min(as.character(getRversion()),"3.5.3")))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <-
 data.frame(
   w=sample(1:4000, size = n, replace = TRUE),
   v=runif(n = n, 0, 10000)
 )

#Run with C++ code.
knapsack_dynamic(knapsack_objects[1:8,], W = 3500, fast = TRUE)

#Run without C++ code.
knapsack_dynamic(knapsack_objects[1:8,], W = 3500, fast = FALSE)
}
